refactoring, fixed teams sorting
